+++ scripts/show-task.js
@@ -10,7 +10,8 @@ import toggleButton from "/scripts/factory/toggleButton.js";
 import onDeleteTask from "/scripts/delete-task.js";
 import onTaskComplete from "/scripts/done-task.js";
 import onTaskEdit from "/scripts/edit-task.js";
-import disableEditing from "/scripts/factory/disableEditing.js";
+import createTask from "/scripts/factory/createTask.js";
+import createToolbar from "/scripts/factory/createToolbar.js";
 
 function showTask(
   inputElement,
@@ -20,38 +21,17 @@ function showTask(
   tasks,
   id
 ) {
-  const task = document.createElement("div");
-  task.classList.add("task");
+  const [task_content, task, textElement] = createTask(inputElement);
 
-  const task_content = document.createElement("div");
-  task_content.classList.add("content");
-
-  const textElement = document.createElement("p");
-  textElement.classList.add("text");
-  textElement.innerText = inputElement.value;
-  disableEditing(inputElement, false);
-
-  task_content.append(textElement);
-
-  const toolbar = document.createElement("div");
-  toolbar.classList.add("actions");
   const date = document.createElement(
     "p",
     "Created At: " + formattedDate,
     true
   );
   const saveButton = createButton(HIDE_CLASS, "Save", "Save the task");
-  const doneButton = createButton(
-    "done",
-    DONE_ICON,
-    "Mark task as done"
-  );
+  const doneButton = createButton("done", DONE_ICON, "Mark task as done");
   const editButton = createButton("edit", EDIT_ICON, "Edit task");
-  const deleteButton = createButton(
-    "delete",
-    DELETE_ICON,
-    "Delete task"
-  );
+  const deleteButton = createButton("delete", DELETE_ICON, "Delete task");
   const saveAndDoneButton = createButton(
     HIDE_CLASS,
     DONE_ICON,
@@ -73,10 +53,11 @@ function showTask(
     revertButton,
   ];
 
-  toolbar.append(date, ...buttons);
+  const toolbar = createToolbar(date, buttons);
   task_content.append(toolbar);
   task.append(task_content);
   taskListElement.prepend(task);
+
   onDeleteTask(deleteButton, task, taskListElement, tasks, id);
   onTaskComplete(doneButton, editButton, textElement, toolbar, tasks, id);
   onTaskEdit(buttons, textElement, toolbar, tasks, id);
